+++ pymc_forecast/__init__.py
@@ -52,6 +52,8 @@ from pymc_forecast.metrics import (
 )
 from pymc_forecast.model import (
     FORECAST_VAR,
+    MU_FORECAST_VAR,
+    MU_VAR,
     OBS_VAR,
     ForecastingModel,
     Horizon,
@@ -76,6 +78,8 @@ __all__ = [
     "DRAW_DIM",
     "FORECAST_VAR",
     "FUTURE_DIM",
+    "MU_FORECAST_VAR",
+    "MU_VAR",
     "OBS_VAR",
     "SAMPLE_DIMS",
     "TIME_DIM",